don't override download options
Unfortunately, R sets '-s -S' in the curl command line when downloading
something with 'download.file(..., quiet = TRUE)', which means that
errors are always emitted to the console. This occurs quite commonly
when attempting to query available packages, since certain package
database files may not be available for some repositories.