Refactor prompting
The major user facing is change the default value of `default` for `proceed()` from `FALSE` to `TRUE`, so the interactive default is equivalent to the non-interactive behaviour.

I then refactored `proceed()` to include a `prompt` argument: I think this makes it more clear what happens when `prompt` is `FALSE`.

While updating the code to use the new `proceed()` behaviour I realised that there many repeated calls that aborted if the user said `n` at the prompt. So I introduced a new `check_can_proceed()` function that extracts out the repeated code.

And then while updating the code to use `check_can_proceed()` I realised that (unsurprisingly) most calls to `check_can_proceed()` came up after a message that was gated by a check for `prompt || renv_verbose()`. So I also added a `prompt` argument to `renv_verbose()` because I think this makes the connection between the two functions more clear.

Fixes #1240