Eliminate use of %||% with environments
This is the first step in moving `%||%` towards an implementation that's consistent with other packages, i.e. what `%??%` is currently. After a few abortive attempts, I figured out the best way to do this is to slowly move usage of `%||%` to `%??%`, since that makes it stricter.

This first PR eliminates the use of `%||%` with environments by using argument defaults more in `scope.R`, and using `%??%` for environment caching elsewhere.

This includes a temporary warning in `%||%` and a temporary elimination of `warn = 1`. If you think this first step is sound, I'll remove these changes prior to merging.