Identify all uses of %||% for non-NULL 0-length vectors
Next step in the standardization of `%||%` semantics. After this `%||%` can be redefined as `function(x, y) if (is.null(x)) y else x` and `%??%` can be removed.  I'm not sure what to call the new function that's used specifically for 0 length vectors.

I'll revert the change to options and to `%||%` prior to merging.