Support GitHub sha as version (#1327)
Instead of installing from a specific version + tag (both of which have be updated on every commit), use the commit sha extracted from the `DESCRIPTION`. Once we've convinced ourselves this is a solid approach, we'll be able to remove the code for updating the version and creating a new tag on each commit to main.