Only report slow dependency discovery for implicit type (#1380)
This was motivated by #1198 where in one of the examples I noticed:

```
NOTE: Dependency discovery took 11 seconds during snapshot.
Consider using .renvignore to ignore files -- see `?dependencies` for more information.
```

That occurs because the current code includes time spent waiting for the prompt. That seemed like a potentially confusing red herring to me , so I moved the time tracking to `renv_snapshot_filter_implicit()`. I think this is correct as (a) it's the only type that's likely to be slow and (b) it's the only type affected by `.renvignore`.