Cache dependencies per file (#1466)
This takes the run-time of `renv::dependencies()` on renv its from 3s down to 0.1s, where ~75% of the remaining time is from `file.info()`.

There is one possible drawback to the current implementation. Because problems are emitted as a side-effect, they're not replaced by the cache so you'll only see them the first time the file is processed. But OTOH, it's nice to only be warned about problems once per session, rather than every time.