Update docker vignette with section on multi-stage builds (#1127)
When images must be built multiple times (e.g., while debugging code),
Docker's caching system is not efficiently used to cache the project
library. Multi-stage builds allow the library cache to be copied
across build stages resulting in significant time savings.

A section is added to the docker vignette describing this process. Minor
reformatting of the 'Dynamically Provisioning R Libraries with renv'
section are also included to increase cohesion of the document.

Co-authored-by: dbenkes <dbenkes@bioc02wg0k1hx8f.wireless.emory.edu>
Co-authored-by: David Benkeser <dbenkeser@gmail.com>