watchdog: fix path to renv's Meta/package.rds
renv_watchdog_start_impl() specifies the library so that the
subprocess can find renv.  If renv appears to be loaded by devtools,
it falls back to setting the library to the result of
renv_libpaths_default().

As of f05df6748 (fix bad refactor; tweaks, 2023-08-07), a missing
Meta/package.rds is taken as an indication that renv was loaded by
devtools.  However, this check incorrectly looks in the _parent_
directory of the renv package.

As a result, f05df6748 resurfaced the problem resolved by 43a3fcdbf
(set library path when loading renv in watchdog process (#1618),
2023-08-01).  The launched process fails 1) if renv is not in the
renv_libpaths_default() paths or 2) if those paths contain an older
renv without the required watchdog functionality.

Point the check to the correct location.