Reattempt package install by pak upon subprocess error
After installing pak by pak, using pak within renv to install
new packages (whether from the lock file or not) leads to
the "Subprocess is busy or cannot start error". However, this
only occurs upon the first attempted install - subsequent installs
work fine.

So, we reattempt the install if the first attempt fails due to a
subprocess error. Other errors are not reattempted (e.g. can't
find package from the remote). Up to three reattempts are attempted,
after which the subprocess error is returned if it persisted.

This is implemented by a new function `renv_pak_retry`, which is
called by `renv_pak_restore` and `renv_pak_restore` (which
themselves are called by `restore`, `update` and `install`). The
function takes the pak install call and the calling
environment as arguments, and evaluates the call in the calling
environment as many times as required up until three times.

I'm sure could definitely replace `env <- new.env()` with just
passing `parent.frame(n = <something>` directly to `renv_pak_retry`,
but I didn't get it right and so left it.